+++ tests/Humanizer.Tests/WordsToNumberLongTests.cs
@@ -63,3 +63,5 @@ public class WordsToNumberLongTests_GB
         Assert.Null(unrecognizedWord);
     }
 }
+
+